+++ DxImageLoader/interface.h
@@ -31,7 +31,7 @@ EXPORT(void) image_info_mipmap(int id, int mipmap, int& width, int& height, int&
 
 /// \brief get mipmap bytes
 /// \return mipmap data. Can also be used to write mipmap data
-EXPORT(unsigned char*) image_get_mipmap(int id, int layer, int mipmap, uint32_t& size);
+EXPORT(unsigned char*) image_get_mipmap(int id, int layer, int mipmap, uint64_t& size);
 
 /// \brief saves the image with the given format and extension
 /// \param id valid image id
